app/backend/requirements.txt
Django>=5.1
djangorestframework>=3.15
djangorestframework-simplejwt>=5.5
psycopg[binary]>=3.2
django-redis>=5.4
python-dotenv>=1.0
drf-spectacular>=0.27
pillow>=10.0
pytest==8.3.3
pytest-django==4.9.0
pytest-sugar==1.0.0
pytest-cov==6.0.0
